+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 Any changes that haven't been included in a published version will be listed here.
 
+- Added `node.reactions = ...` to the list of banned property setters under `ban-deprecated-sync-prop-setters`. Use `node.setReactionsAsync` instead.
+
 ## 0.13.0
 
 - Updated `ban-deprecated-sync-prop-getters` rule to allow _assignment_ to properties. This allows for statements such as `instance.mainComponent = ...`.